Reorganize zosconsole params objects into input package and finished javadoc for current zosconsole package
