+++ src/main/javadoc/zosconsole/zosmf/ConsoleParms.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>zosconsole.zosmf.ConsoleParams</li>
+<li>zosconsole.input.ConsoleParams</li>
 </ul>
 </li>
 </ul>
