+++ src/main/javadoc/zosconsole/zosmf/IssueParams.html
@@ -123,10 +123,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li><a href="ConsoleParams.html" title="class in zosconsole.zosmf">zosconsole.zosmf.ConsoleParams</a></li>
+<li><a href="ConsoleParams.html" title="class in zosconsole.zosmf">zosconsole.input.ConsoleParams</a></li>
 <li>
 <ul class="inheritance">
-<li>zosconsole.zosmf.IssueParams</li>
+<li>zosconsole.input.IssueParams</li>
 </ul>
 </li>
 </ul>
